new changes

+++ printf_dec.c
@@ -8,7 +8,7 @@
  */
 int printf_dec(va_list d)
 {
-	int dig, count = 1, len = 0, temp, x;
+	long int dig, count = 1, len = 0, temp, x;
 
 	dig = va_arg(d, int);
 	temp = dig;
